Ajusta tela de perfil do usuario logado
