+++ app/screens/perfilPaciente/PerfilPaciente.tsx
@@ -1,11 +1,10 @@
-import React, { useState, useRef } from 'react';
+import React, { useState, useRef, useEffect } from 'react';
 import {
   View,
   Text,
   TextInput,
   Image,
   TouchableOpacity,
-  StyleSheet,
   Alert,
   Animated,
   Modal,
@@ -13,24 +12,34 @@ import {
 } from 'react-native';
 import { Feather } from '@expo/vector-icons';
 import Sidemenu from '../Sidemenu';
-import perfilPhoto from '../../../icons/perfilPhoto.jpg'; // Substitua pelo caminho correto da imagem
+import styles from '@/styles/PerfilPacienteStyle';
+import { useUsuarioLogado } from '@/hooks/UsuarioLogadoProvider ';
+import { PacienteModel } from '@/constants/models/PacienteModel';
+import LogoutButton from '@/common/LogoutButton';
+import TrocarSenhaButton from '@/common/TrocarSenhaButton';
+import PsicologoProvider from '@/app/provider/PsicologoProvider';
+import { useLoading } from '@/hooks/LoadingContext';
 
 export default function ProfileScreen() {
-  const [code, setCode] = useState('');
-  const [professional, setProfessional] = useState('Julie Smith, Psychologist');
+  const { usuarioLogado } = useUsuarioLogado();
+  const {showLoading , hideLoading} = useLoading();
 
-  const [email, setEmail] = useState('johndoe@example.com');
-  const [phone, setPhone] = useState('(123) 456-7890');
+  const pacienteLogado = usuarioLogado.usuarioLogadoData as PacienteModel
+
+  const [professional, setProfessional] = useState('');
+
+  const [email, setEmail] = useState(pacienteLogado.nome);
+  const [phone, setPhone] = useState(pacienteLogado.telefone);
 
   const [isModalVisible, setModalVisible] = useState(false);
   const [fieldBeingEdited, setFieldBeingEdited] = useState<'email' | 'phone' | null>(null);
   const [tempValue, setTempValue] = useState('');
 
-  const scalePhoto = useRef(new Animated.Value(1)).current;
+  const uri = `https://randomuser.me/api/portraits/men/${pacienteLogado.idPaciente}.jpg`
   const scaleEmail = useRef(new Animated.Value(1)).current;
   const scalePhone = useRef(new Animated.Value(1)).current;
 
-  function animateIcon(scaleRef: Animated.Value) {
+  const animateIcon = (scaleRef: Animated.Value) => {
     Animated.sequence([
       Animated.timing(scaleRef, {
         toValue: 1.2,
@@ -45,12 +54,8 @@ export default function ProfileScreen() {
     ]).start();
   }
 
-  function handleEditPhoto() {
-    animateIcon(scalePhoto);
-    Alert.alert('Editar Foto', 'Funcionalidade para editar a foto.');
-  }
 
-  function openEditModal(field: 'email' | 'phone') {
+  const openEditModal = (field: 'email' | 'phone') => {
     if (field === 'email') animateIcon(scaleEmail);
     else animateIcon(scalePhone);
 
@@ -59,7 +64,7 @@ export default function ProfileScreen() {
     setModalVisible(true);
   }
 
-  function handleSave() {
+  const handleSave = () => {
     if (fieldBeingEdited === 'email') {
       setEmail(tempValue);
     } else if (fieldBeingEdited === 'phone') {
@@ -68,104 +73,52 @@ export default function ProfileScreen() {
     setModalVisible(false);
   }
 
-  function handleCancel() {
+  const handleCancel = () =>{
     setModalVisible(false);
   }
 
-  function handleChangePassword() {
+  const handleChangePassword = () =>{
     Alert.alert('Trocar senha', 'Funcionalidade para trocar a senha.');
   }
 
-  function handleRemoveProfessional() {
-    Alert.alert('Desvincular', 'Profissional removido.');
-    setProfessional('');
-  }
-
-  function handleConnectProfessional() {
-    if (code.trim()) {
-      setProfessional('Novo Profissional (exemplo)');
-      setCode('');
-      Alert.alert('Sucesso', 'Profissional vinculado.');
-    } else {
-      Alert.alert('Erro', 'Digite um código válido.');
-    }
-  }
-
-  function handleLogout() {
-    Alert.alert('Sair', 'Você saiu do app.');
-  }
-
-  function handleEditEmail() {
-    Alert.alert('Editar Email', 'Funcionalidade para editar o email.');
-  }
-
-  function handleEditPhone() {
-    Alert.alert('Editar Telefone', 'Funcionalidade para editar o telefone.');
-  }
+  useEffect(()=>{
+    showLoading()
+    PsicologoProvider.obterInformacoesPsicologoPorPaciente(pacienteLogado.fk_idProfissional).then((psicologoResponse)=>{
+      const nomePsicologoResponsavel = psicologoResponse.nome
+      setProfessional(nomePsicologoResponsavel)
+    }).catch((erro)=>{
+      console.error("Falha ao obter informacoes do psicologo responsavel, ", erro)
+    }).finally(()=>{
+      hideLoading()
+    })
+  },[usuarioLogado])
 
   return (
     <View style={styles.container}>
       <View style={styles.profileImageContainer}>
-        <Image source={{ uri: perfilPhoto }} style={styles.profileImage} />
-
-        <Animated.View style={[styles.editIcon, { transform: [{ scale: scalePhoto }] }]}>
-          <TouchableOpacity onPress={handleEditPhoto}>
-            <Feather name="edit-2" size={18} color="#333" />
-          </TouchableOpacity>
-        </Animated.View>
+        <Image source={{ uri }} style={styles.profileImage} />
       </View>
-
       <Text style={styles.name}>John Doe</Text>
-
       <View style={styles.infoRow}>
         <Text style={styles.info}>{email}</Text>
-        <Animated.View style={{ transform: [{ scale: scaleEmail }] }}>
-          <TouchableOpacity onPress={() => openEditModal('email')}>
-            <Feather name="edit-2" size={16} color="#333" style={styles.smallEditIcon} />
-          </TouchableOpacity>
-        </Animated.View>
       </View>
-
       <View style={styles.infoRow}>
         <Text style={styles.info}>{phone}</Text>
-        <Animated.View style={{ transform: [{ scale: scalePhone }] }}>
+        {phone && <Animated.View style={{ transform: [{ scale: scalePhone }] }}>
           <TouchableOpacity onPress={() => openEditModal('phone')}>
             <Feather name="edit-2" size={16} color="#333" style={styles.smallEditIcon} />
           </TouchableOpacity>
-        </Animated.View>
+        </Animated.View>}
       </View>
-
-      <TouchableOpacity style={styles.button} onPress={handleChangePassword}>
-        <Text style={styles.buttonText}>Trocar Senha</Text>
-      </TouchableOpacity>
-
       <View style={styles.section}>
         <Text style={styles.sectionTitle}>Profissional</Text>
-        {professional ? (
-          <View style={styles.professionalContainer}>
+        <View style={styles.professionalContainer}>
             <Text style={styles.professionalText}>{professional}</Text>
-            <TouchableOpacity style={styles.removeButton} onPress={handleRemoveProfessional}>
-              <Text style={styles.removeButtonText}>Remover</Text>
-            </TouchableOpacity>
-          </View>
-        ) : (
-          <View style={styles.connectContainer}>
-            <TextInput
-              style={styles.input}
-              placeholder="Digite o código"
-              value={code}
-              onChangeText={setCode}
-            />
-            <TouchableOpacity style={styles.connectButton} onPress={handleConnectProfessional}>
-              <Text style={styles.connectButtonText}>Conectar</Text>
-            </TouchableOpacity>
-          </View>
-        )}
+        </View>
       </View>
-
-      <TouchableOpacity style={styles.logoutButton} onPress={handleLogout}>
-        <Text style={styles.logoutButtonText}>Sair do App</Text>
-      </TouchableOpacity>
+      <LogoutButton></LogoutButton>
+      
+      <TrocarSenhaButton handleChangePassword={handleChangePassword}></TrocarSenhaButton>
 
       <Modal
         animationType="slide"
@@ -175,9 +128,6 @@ export default function ProfileScreen() {
       >
         <View style={styles.modalBackground}>
           <View style={styles.modalContainer}>
-            <Text style={styles.modalTitle}>
-              Editar {fieldBeingEdited === 'email' ? 'Email' : 'Telefone'}
-            </Text>
             <TextInput
               style={styles.modalInput}
               value={tempValue}
@@ -196,7 +146,6 @@ export default function ProfileScreen() {
           </View>
         </View>
       </Modal>
-
       <View style={styles.sideMenuContainer}>
         <Sidemenu />
       </View>
@@ -204,159 +153,3 @@ export default function ProfileScreen() {
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    padding: 20,
-  },
-  profileImageContainer: {
-    position: 'relative',
-    marginTop: 20,
-    width: 120,
-    height: 120,
-    backgroundColor: '#e0e0e0',
-    alignItems: 'center',
-  },
-  profileImage: {
-    width: 120,
-    height: 120,
-    borderWidth: 2,
-    borderRadius: 50,
-    backgroundColor: 'gray', // Adicione isso temporariamente
-  },
-  editIcon: {
-    position: 'absolute',
-    bottom: 0,
-    right: 0,
-    backgroundColor: '#fff',
-    padding: 5,
-    borderRadius: 20,
-    elevation: 5,
-  },
-  name: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    marginTop: 10,
-  },
-  infoRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginTop: 10,
-  },
-  info: {
-    fontSize: 16,
-  },
-  smallEditIcon: {
-    marginLeft: 8,
-  },
-  modalBackground: {
-    flex: 1,
-    backgroundColor: 'rgba(0,0,0,0.5)',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  modalContainer: {
-    width: '80%',
-    backgroundColor: '#fff',
-    padding: 20,
-    borderRadius: 10,
-    elevation: 10,
-  },
-  modalTitle: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    marginBottom: 10,
-  },
-  modalInput: {
-    borderWidth: 1,
-    borderColor: '#ccc',
-    borderRadius: 5,
-    padding: 10,
-    marginBottom: 20,
-  },
-  modalButtons: {
-    flexDirection: 'row',
-    justifyContent: 'flex-end',
-  },
-  modalButtonCancel: {
-    marginRight: 10,
-  },
-  modalButtonSave: {},
-  modalButtonText: {
-    color: '#007bff',
-    fontWeight: 'bold',
-  },
-
-  button: {
-    marginTop: 20,
-    backgroundColor: '#e0e0e0',
-    paddingVertical: 10,
-    paddingHorizontal: 30,
-    borderRadius: 8,
-  },
-  buttonText: {
-    fontSize: 16,
-    color: '#333',
-  },
-  section: {
-    width: '100%',
-    marginTop: 30,
-  },
-  sectionTitle: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    marginBottom: 10,
-  },
-  professionalContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-  },
-  professionalText: {
-    fontSize: 16,
-  },
-  removeButton: {
-    backgroundColor: '#f44336',
-    padding: 5,
-    borderRadius: 5,
-  },
-  removeButtonText: {
-    color: '#fff',
-  },
-  connectContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  input: {
-    borderWidth: 1,
-    padding: 8,
-    borderRadius: 5,
-    marginRight: 10,
-    flex: 1,
-  },
-  connectButton: {
-    backgroundColor: '#007bff',
-    padding: 8,
-    borderRadius: 5,
-  },
-  connectButtonText: {
-    color: '#fff',
-  },
-  logoutButton: {
-    marginTop: 20,
-    backgroundColor: '#f44336',
-    paddingVertical: 10,
-    paddingHorizontal: 30,
-    borderRadius: 8,
-  },
-  logoutButtonText: {
-    color: '#fff',
-  },
-  sideMenuContainer: {
-    position: 'absolute',
-    bottom: 0,
-    left: 0,
-    right: 0,
-  },
-});
