+++ common/TrocarSenhaButton.tsx
@@ -1,0 +1,32 @@
+import React from 'react';
+import { TouchableOpacity, View, Text, StyleSheet } from 'react-native';
+
+type Props = {
+  handleChangePassword: () => void;
+};
+
+const TrocarSenhaButton: React.FC<Props> = ({ handleChangePassword }) => {
+  return (
+    <View>
+      <TouchableOpacity style={styles.button} onPress={handleChangePassword}>
+        <Text style={styles.buttonText}>Trocar Senha</Text>
+      </TouchableOpacity>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  button: {
+    marginTop: 20,
+    backgroundColor: '#e0e0e0',
+    paddingVertical: 10,
+    paddingHorizontal: 30,
+    borderRadius: 8,
+  },
+  buttonText: {
+    fontSize: 16,
+    color: '#333',
+  },
+});
+
+export default TrocarSenhaButton;
