+++ styles/LogoutButtonStyle.ts
@@ -1,0 +1,20 @@
+import { StyleSheet } from 'react-native';
+
+
+const styles = StyleSheet.create({
+
+    logoutButton: {
+      marginTop: 20,
+      backgroundColor: '#f44336',
+      paddingVertical: 10,
+      paddingHorizontal: 30,
+      borderRadius: 8,
+    },
+  
+    logoutButtonText: {
+      color: '#fff',
+    }
+  });
+  
+  export default styles;
+
